1-click installer script
Hello. First of all, I want to thank you for the AHI, it works well. However, I thought that the installation process could've been more streamlined. For this reason, I have made a script that automatically downloads, extracts, installs, and copies the latest versions of all the necessary packages. It cleans up after itself as well but leaves the AHI folder since it contains useful examples and tools (like `Monitor.ahk`). It also prompts to restart the computer at the end. It does not have any error checking routines, but it enforces the admin & version requirements to eliminate the most common issues. I'm posting it here in case someone might find it useful or convenient, under the CC0 license.